pyproject.toml
[project]
name = "nonebot-plugin-yesman"
version = "1.2.0"
description = "Yesman for NoneBot2"
authors = [
    {name = "euds63", email = "erica23187@gmail.com"},
]
license = "MIT"
dependencies = ["nonebot2>=2.0.0b2", "nonebot-adapter-onebot>=2.0.0b1"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/EuDs63/nonebot-plugin-yesman"
Repository = "https://github.com/EuDs63/nonebot-plugin-yesman"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
